+++ esbuild.js
@@ -1,4 +1,0 @@
-import {buildScripts, buildStyles} from '../yii2-skeleton/esbuild.config.js';
-
-await buildScripts();
-await buildStyles();