+++ src/Modules/Admin/Widgets/Forms/LocationProviderIdField.php
@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Hirtz\Location\Modules\Admin\Widgets\Forms;
 
-use Hirtz\Location\Modules\Admin\Assets\AutocompleteAssetBundle;
 use Hirtz\Location\Modules\Admin\Module;
 use Hirtz\Location\Modules\ModuleTrait;
-use Hirtz\Skeleton\Widgets\Forms\Fields\InputField;
+use Hirtz\Skeleton\Widgets\Forms\Fields\AutocompleteField;
 use Override;
 use Yii;
 
-class LocationProviderIdField extends InputField
+class LocationProviderIdField extends AutocompleteField
 {
     use ModuleTrait;
 
@@ -20,31 +19,18 @@ class LocationProviderIdField extends InputField
     {
         $this->property ??= 'provider_id';
 
-        $this->attributes['autocomplete'] ??= 'off';
-
         if (!$this->model->{$this->property}) {
             $this->attributes['placeholder'] ??= Yii::t('location', 'LOCATION_PROVIDER_ID_SEARCH');
             $this->attributes['type'] ??= 'search';
         }
 
-
         /** @var Module $module */
         $module = Yii::$app->getModule('admin')->getModule('location');
 
         if ($module->getAutocomplete()) {
-            $this->registerAutocompleteClientScript();
+            $this->url ??= ['/admin/location/location/autocomplete'];
         }
 
         parent::configure();
-    }
-
-    protected function registerAutocompleteClientScript(): void
-    {
-        $module = AutocompleteAssetBundle::register($this->view);
-
-        $this->view->registerJsModule("$module->baseUrl/$module->filename", [
-            '#' . $this->getId(),
-            Yii::$app->getUrlManager()->createUrl(['/admin/location/location/autocomplete']),
-        ]);
     }
 }